backend/requirements.txt
# Web フレームワーク
fastapi==0.115.5
uvicorn[standard]==0.32.1

# データベース
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# 設定・バリデーション
pydantic[email]==2.10.3
pydantic-settings==2.6.1

# 認証
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.19

# Google OAuth
authlib==1.3.2
httpx==0.28.1

# ファイル出力（スキルシート）
openpyxl==3.1.5
reportlab==4.2.5

# スケジューラー（通知・アラート）
apscheduler==3.10.4

# AI検索
anthropic==0.40.0

# テスト
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
aiosqlite==0.20.0
